Readability improvements Make hooks directory a module Add extra check for pagination display Rename hook to useContent to keep name consistency
